Add 'isomorphic-fetch' import to index
This ensures that 'isomorphic-fetch' is imported before anything else.

Previously 'isomorphic-fetch' was imported at the beginning of each
separate module that used 'fetch', which was needlessly verbose and
easy to forget to do (it was missing from one module already).

This would normally break the tests, as the tests don't import modules
via the main index file. This was resolved by adding a `setupTests`
file that imports `isomorphic-fetch` before running any tests.