Handle unsuccessful fetch (#157)
* Add 'isomorphic-fetch' import to index

This ensures that 'isomorphic-fetch' is imported before anything else.

Previously 'isomorphic-fetch' was imported at the beginning of each
separate module that used 'fetch', which was needlessly verbose and
easy to forget to do (it was missing from one module already).

This would normally break the tests, as the tests don't import modules
via the main index file. This was resolved by adding a `setupTests`
file that imports `isomorphic-fetch` before running any tests.

* Handle unsuccessful fetch attempts

The `fetch` API will throw an exception if a network error is
encountered during a `fetch` invocation, but an unsuccessful response
will not. We should never assume that the response from `fetch` is
successful without checking first. We can't even assume that
successfully calling `.json()` on a request implies that it succeeded,
because failed responses can pass information in the response body as
well.

A `successfulFetch` utility method has been added that verifies the
fetch response was successful, and throws an error if it was not.
This method is now used for the other two `fetch` utility methods, and
each instance of `fetch` has been updated to use one of these methods.

New tests have been added for `successfulFetch`. `fetch-mock` was
updated to make it easier to specify the status of responses from calls
to the mock fetch instance.