Prevent deadlocks resulting from errors thrown before releasing mutex
A `try ... finally` block has been added to most places where a mutex
is used, to ensure that any error thrown before the lock has been
released will not result in a deadlock.

The `finally` block will execute after the `try` block, but it will
not interfere with any returned values or thrown errors (so long as
nothing is thrown or returned within the `finally` block).