Implement redesigned controller API
This represents a prototype implementation of a new controller API
design. The motivations for this are written about in this document:
https://docs.google.com/document/d/1DjtFfNB0JuziijGqs4h3vW8PL8zCI6zuaovqvSHbJ2o/edit#heading=h.mtvpfhpuzjof

A new base controller has been introduced (`base-controller-v2.ts`),
and the `CurrencyRateController` has been updated to use it. All other
controllers remain unchanged.

Note that the `CurrencyRateController` no longer works correctly with
the `ComposableController` - I didn't think that was worth the effort
to get working.