Remove unused dependencies
One dependency and two dev dependencies were unused, and have been
removed.

`percentile` has been in this repository for a very long time - it's
unclear when it became unused, or if it was ever used.

`jsdom` was no longer used, and the types for `xtend` haven't been
needed since `xtend` was removed from the project.