Add Dependabot config (#343)
This config will direct Dependabot to update the minimum version of any
internal dependencies to match the latest publish version.

Since this is a library, there's no particular need for us to update
our dependencies when new versions are released that are within the
current range. However, it is still convenient. We often want to
require updates because they bring in bug fixes or other improvements,
and this makes updating lockfiles in each product easier (as Yarn will
keep the old version of the library in the lockfile if it can, which
requires manual intervention to prevent).