Add `babel-runtime` dependency required by `ethjs-query` (#341)
The version of `ethjs-query` we're using has an unlisted dependency
upon `babel-runtime`, and will crash if it's not present. Currently
it's working because `babel-runtime` is being brought in by another
dependency (`fetch-mock`), but that dependency is removed in another
PR (#340).

Until we can migrate away from using this version of `ethjs-query`,
`babel-runtime` has been added as an explicit dependency to ensure it
is present.

`ethjs-query` is a transitive dependency of `eth-method-registry` via
`ethjs`. There is a PR open on `eth-method-registry` to replace `ethjs`
altogether [1]. Once those are merged and released, we can update
`eth-method-registry` here and remove this `babel-runtime` dependency.

[1]: https://github.com/MetaMask/eth-method-registry/pull/29