Prevent `ApprovalController` counting mismatch (#356)
The `ApprovalController` was not great at counting. If you tried to
delete an approval that didn't exist, or if you tried to add an
approval that already existed, it would miscount the current number
of pending approvals.

Validation was in place already to prevent both scenarios, so in
practice this shouldn't be possible without calling private functions.
But still, preventing it entirely was trivial, so here it is.