Improve anonymizer return type
The anonymizer function can now return a _recursive partial_ of the
input type. So any properties can be omitted in the state property
directly or in any nested object.

Note that for primitive values and arrays, the return type must still
match the state. Similarly, any objects still need to be returned as
objects, though they're allowed to have a few less properties.