Constrain BaseController state to be valid JSON
The BaseController state is now constrained to valid JSON. Anything
that can't be serialized (e.g. functions) or gets mutated during
serialization (e.g. `undefined` gets converted to `null`) is
disallowed.

This should prevent an entire class of bugs resulting from unexpected
changes when serializing state as JSON. For example, we serialize state
when sending it over `postMessage`, and when persisting the state.