Add BaseControllerV2 state metadata (#371)
State metadata has been added to the new BaseController constructor as
a required constructor parameter. The metadata describes how to derive
the state that should be persisted, and how to derive an 'anonymized'
representation of the controller state.

The metadata describes top-level properties only, but it allows you to
define a function to derive the anonymized or persistent state for each
property. The only requirement of this derivation function is that the
output is also valid JSON.

This is part of the controller redesign (#337).