Add restricted controller messenger
A method has been added to the controller messenger that returns a
_restricted_ instance of the controller messenger. This instance is
restricted to registering actions and events within a particular
namespace, typically being the name of the controller. Access to event
subscriptions and calling actions is also restricted based on the given
allowlist for each.

This is meant to address feedback left on the Controller Messaging
System proposal[1].

[1]: https://www.notion.so/Controller-Messaging-System-617efb02b9e54bd0a0b0e44c6f776d85