Fix `format` commands
The `format` commands were only applying to files one directory down.
Files in subdirectories were not formatted. This is because the file
pattern used in both commands was unquoted, and so was expanded by
Bash. Bash doesn't support the globstar operator by default, so it
only patched the directory contents at one level.

The file pattern has been quoted, which ensures that it gets passed
to `prettier` intact rather than expanded. `prettier` does support the
globstar operator, so it now functions as we'd expect.