Add controller messaging system (#377)
Adds a controller messaging system, which will be used to facilitate
inter-controller communication. The controller messenger acts as a
message broker, passing actions and events back and forth between
controllers. It is _fully type safe_.

This idea was described in the Controller Messaging System proposal[1].

This controller messenger doesn't yet support attenuation, so it's not
yet possible to restrict which actions and events are interacted with.
That will be coming in a later PR.

[1]: https://www.notion.so/Controller-Messaging-System-617efb02b9e54bd0a0b0e44c6f776d85

* Remove `subscribe` and `unsubscribe` BaseController methods

The base controller no longer exposes methods to subscribe and
unsubscribe to state changes. Instead all subscriptions and
unsubscriptions are done with the controller messaging system.