Improve asset metadata type
The asset metadata type has been updated to conditionally require the
`error` field for errors, and disallow it for non-errors. This allows
us to remove the non-null assertion used in the 'finished' callback,
as now TypeScript can infer that the `error` field must be present if
the status is `failed`.