Use consistent type for token balances
We have been using inconsistent types for token balances. We use a `BN`
in most places, but if we fail to find a balance we set it to the
number `0`.

We now use `BN` in all cases; `new BN(0)` is the fallback now instead
of `0`.

This error was made possible because of various type errors surrounding
`BN` and these balances. `typeof BN` was used as the type instead of
`BN`, because `BN` was imported using `require` and TypeScript was
treating it as a value rather than a class, and unhelpfully suggested
using `typeof BN`. This resulted in the type being `any` in effect.

We now import `BN` using `import`, and use `BN` as the type.

This highlighted another problem as well: the return type of the
`getBalancesInSingleCall` function in `AssetsContractController` was
incorrect. It claimed to return a string balance, but in fact it
returned a `BN` balance. The type has been corrected. Fortunately we
were not actually _using_ the balance anywhere (I checked mobile as
well). We only use the keys of the return type. The main utility of the
method is that it automatically omits any tokens with zero balance.

Lastly, the type corrections also highlighted that we had an export not
covered by unit tests. The type `BN` was being re-exported, presumably
in response to a TypeScript error about the use of a private type. This
is another case of TypeScript giving bad advice when it doesn't have
proper type information. The unused export has been removed. I
confirmed that it was unused in this repo as well as in mobile.