Fix `signTransaction` transaction parameter type
The `signTransaction` method of the KeyringController had specified the
wrong type for the `transaction` parameter. It mistakenly declared it
as being of type `Transaction`, the same type we use in our Transaction
controller.

Instead the type it expects is an instance of `ethereumjs-tx`. This
package doesn't have type definitions though, so for now I've changed
it to `unknown`, and updated the JSDoc comment to explain what it
should be.