Migrate mock encryptor to TypeScript (#399)
The mock encryptor we use for the KeyringController tests has been
migrated to TypeScript. It has also been converted from an object
literal to a class, as it seemed a bit more conventional that way. The
`encrypt` function is also no longer a sinon stub - the stub
functionality was not used at all.