Only apply TypeScript lint rules to `.ts` files (#396)
The TypeScript lint rules were being applied to all files covered by
the linter, including JavaScript files. Surprisingly this hasn't been
much of an issue (I guess most of them are compatible with JavaScript
too?), but it became an obstacle when trying to re-enable the rule
`@typescript-eslint/no-shadow`.

TypeScript lint rules are now only applied to TypeScript files.

A new dependency was required to ensure we could import TypeScript
modules from JavaScript modules. We do this in one place at the moment,
in the JavaScript tests for the ApprovalController. Apparently the
ESLint `import` plugin doesn't know how to resolve TypeScript modules
without a custom resolver. We hadn't encountered this problem
previously because `import/no-unresolved` was disabled in our
TypeScript ESLint config, so it didn't warn us about the dependency
being unresolved.

The `ecmaVersion` was also updated to be 2018 for JavaScript files, to
reflect the language features we currently rely upon (we use object
spread in one place, which was introduced in ES2018).