Improve TransactionMeta type
The TransactionMeta type now uses an enum rather than a string for the
`status` property. This enum also determines whether the `error`
property is present. This allows us to get rid of the non-null
assertion used previously to reference this `error` property, as the
type system is now capable of determining that the property exists all
on its own.

A default error was added for any failed transactions from the
Etherscan API, as they didn't have an `error` property beforehand. The
error simply states that the transaction failed, as unfortunately I
don't know that there is any more information available about the
nature of the failure.