Re-enable `accessor-pairs` ESLint rule
The `accessor-pairs` ESLint rule has been re-enabled, which requires
any setter to have a corresponding getter (and vice-versa).

All of our rule violations were for properties that we use only as
setters. In each case a getter was added that throws an error, which
should help prevent accidental use of the getter. The default behaviour
was to return `undefined`, which could lead to silent failures.

Long-term we should stop using setters like this. If we want to allow
changing a property but not accessing, we can use a method. Methods are
generally easier to understand than setters because it's clear to the
consumer that side-effects are possible, and they have more informative
type signatures (e.g. they can return a Promise when they perform some
async operation, so that the caller knows when the operation has
finished).

But replacing the setters would be a breaking change, so that has been
left for another time. A TODO comment has been left to remind us to
replace the setter with a method (e.g. during the migration to the new
base controller).