Re-enable the `@typescript-eslint/no-unused-vars` rule (#407)
The rule `@typescript-eslint/no-unused-vars` has been re-enabled with a
rule adjustment: sibling properties to the rest operator are now
ignored. This is helpful for cases where you want to use the rest
operator to _exclude_ certain properties, which we've done once in the
NetworkController.

A TODO comment has been left about migrating this rule adjustment back
into `@metamask/eslint-config`.