Enable `esModuleInterop` (#404)
This flag makes it easier to work with ES5 dependencies in TypeScript,
and it improves type checking for all dependencies by ensuring
namespace imports aren't called accidentally. It is described in more
detail in the changelog where it was introduced [1].

In particular, see this note:

>Note: The new behavior is added under a flag to avoid unwarranted
>breaks to existing code bases. *We highly recommend applying it both*
>*to new and existing projects*. For existing projects, namespace
>imports (`import * as express from "express"; express();`) will need to
>be converted to default imports
>(`import express from "express"; express();`).

Any namespace imports where the namespace is called have been converted
to default imports. This only applied to two dependencies: `nock` and
`sinon`.

[1]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-7.html#support-for-import-d-from-cjs-from-commonjs-modules-with---esmoduleinterop