Add contract address validation
Contract address validation has been added to the
`getCollectibleTokenURI` method. This validation was already present in
the one place where `getCollectibleTokenURI` is called, so overall this
is basically a refactor with no functional changes, except that the API
of each of these controllers has changed. It's a non-functional change
from mobile's perspective though.

It might seem more appropriate to throw an error if the address isn't
an NFT, rather than returning an empty string. But this is the pre-
existing behaviour, so it has been preserved for now.