Update `selectedAddress` when identities are updated
The `selectedAddress` state is now reset to the first identity if the
current selected address is removed during an update. This is to ensure
the `selectedAddress` is never set to an address that is not present in
state.

This was already being done manually after each invocation of
`updateIdentities` where account removal was anticipated. For mobile
this is mostly a non-functional change, except that now this update
happens in a single state update rather than over the course of two
successive updates.