Restore `BN` export
The `BN` export was removed in #398. I hadn't anticipated at the time
that this would constitute a breaking change. It turns out that this
export was relied upon in mobile in one place.

It has been temporarily restored, along with a comment that suggests
we remove it in the next major version.