Use DI for controller communication rather than context (#387)
We now setup inter-controller communication using dependency injection
rather than the `context` property. Any dependency a controller has is
passed in via a constructor parameter. This was done in preparation for
migrating to BaseControllerV2 and the new controller messaging system -
it's just a temporary solution that will let us migrate controllers one
at a time.

The style of dependency injection here matches the extension. Just as
we do there, we inject state snapshots, means of subscribing to state,
and individual methods rather than entire controllers. This helps to
simplify tests and makes it easier to understand how controllers
interact.