Remove ComposableController generic type parameters
These generic type parameters weren't really used in practice. They
were passed to the `messagingSystem`, but we erased the type of the
messaging system in the one place that we call it, making the more
specific types pointless. You can tell that they weren't useful because
they were never set in tests, and everything still worked.