Add BaseControllerV2 support to ComposableController (#447)
The ComposableController now supports controllers that are based upon
BaseControllerV2. These newer controllers use the controller messaging
system to communicate, so the messenger is now a required constructor
parameter if any BaseControllerV2 controllers are used.

The `name` property in `BaseControllerV2` had to be made public for the
ComposableController to properly construct its state. This seems OK. It
was only kept private initially until a use was found for it.