Simplify BaseControllerV2 tests (#448)
The unit tests for BaseControllerV2 have been updated to be more
concise and (hopefully) slightly easier to read. The restricted
controller setup was moved to a helper function, and common elements
of various types were extracted and re-used.

There were a few cases where I could have used the controller name
variable in a string literal, but chose not to (for example, the
`allowedActions` string literals). This is because of the poor state
of type inference right now for string literal types in TypeScript.
They would get inferred as type `string` each time, which resulted in a
type error. The only way to get it to work was to follow it with
`as '[the string literal type]', which is strictly worse than just
using the string literal in the first place. There are some
improvements planned here in TypeScript v4.3 though, so maybe we'll
get to to deduplicate the controller name in more places soon!