Restore ComposableController compatibility between different BaseControllers (#458)
When adding support for BaseControllerV2 to the ComposableController
in #447, we accidentally removed support for using two different
versions of the BaseController in the same ComposableController. This
is because we used the `instanceof` operator to check whether a
controller was extended from BaseController.

The `instanceof` check has been replaced by a check for the `subscribe`
function. It seems unlikely that a BaseControllerV2 controller will
ever have a `subscribe` function because the controller messenger
will be used for all events.

The BaseControllerV2 class now prevents the use of the `subscribe`
property, ensuring the duck typing strategy used by the
ComposableController to tell old and new controllers apart continues to
work.