Rename `handle to `pollTimeout`
The `handle` private variable has been renamed to better reflect what
it is. Also the type was updated to the more-specific `NodeJS.Timeout`.

It seems a bit odd to call this a "timeout" given that it's for a
repeating interval, but this is the term the Node.js docs use, and I
can't think of a better one.