Improve restricted messenger types for controllers
The `getRestricted` method has been improved to further narrow the
types of the restricted messaging controller. This allows
BaseControllerV2-based controllers to have full type safety for their
usage of the controller messenger.

The BaseController itself still doesn't have a type-safe controller
messenger; that will be dealt with in a later PR (if we ever get it
working).