Add changelog validation to CI
The changelog is now validated during CI. The validation differs
depending on whether the branch is a release branch or not, so the
README was updated to clarify what naming convention we're now using
for release branches.

The `repository` entry in `package.json` was updated to use HTTPS,
because that is the format supported by `auto-changelog`.