Remove uneccessary calls to Collection.set in contoller messenger (#532)
The unrestricted controller messenger implementation contained a couple of unnecessary calls to Collection.set. This PR removes them.

Every removed / changed instance consisted of mutating the object value (actually also collections) of some key in a collection, and then setting the key to the same object value. Mutating the object value of a collection does not affect the collection's reference to said object, so those calls were unnecessary.