Move controller messenger type demonstration to its own test suite (#565)
Previously, we demonstrated how controller messenger generics could be a union of actions and events from different controllers in the currency rate controller tests. Although we have use cases in production that demonstrate this, I think it's worthwhile to keep an example in the tests. This PR moves this type demonstration into the controller messenger test suite.