Update to v9 ESLint config (JSDoc linting)
Our ESLint rules have been updated to v9, which includes new rules to
enforce our JSDoc conventions.

Most of the changes in this PR were a result of running
`yarn lint:fix`. There were a handful of manual changes that seemed
obvious and simple to make.

There should be no functional changes in this PR. All changes are to
comments (aside from the config update and related config changes).