Allow JSDoc descriptions to end with list of links (#617)
When giving descriptions to functions, classes, etc., it can sometimes
be useful to add a list of links at the end which the reader may wish
to visit in order to learn more. For instance:

``` javascript
/*
 * This function does something important. For more information, see:
 *
 * - <https://foo.com>
 * - <https://bar.com>
 */
function foo() {
  // do something
}
```

This commit adds support for the above. It also adds support for
Markdown links with alternate text, e.g. `[some link](https://foo.com)`,
as well as the use of `@link`, e.g. `{@link SomeClass}`.