Pull in networkCongestion along with gas estimates
The Metaswap API now includes a `networkCongestion` property when
requesting EIP-1559-compatible gas fee estimates. This value, which is a
number from 0 to 1, where 0 represents "not congested" and 1 represents
"extremely congested", can be used to communicate the status of the
network to users when they are sending transactions.

This commit ensures that GasFeeController includes `networkCongestion`
as part of the state it persists, no matter if using the Metaswap API to
obtain this information or falling back to `eth_feeHistory`.

This is a part of the EIP-1559 v2 work.