Refactor EIP-1559 v2 fallback code to ease expansion
In future commits we will bring in more gas fee related data points from
the MetaSwap API. When we do this, we will have to implement fallbacks
for all of these data points. Every data point is different and will
require a different strategy and different testing. This commit breaks
up the existing fallback code so this is a lot easier to do.