Pull in remaining EIP-1559 v2 datapoints
Several new properties are now available from the `/suggestedGasFees`
endpoint in the MetaSwap API. This commit updates GasFeeController to
store these properties in state when polling for gas fee estimates and
implements fallback code for the underlying datapoints. This way we can
use these values to display different UI elements to users when they are
sending transactions.

The properties are:

* `historicalBaseFeeRange` — minimum and maximum values for the base
  fees over the past 200 blocks
* `baseFeeTrend` — an indicator for which direction the base fee is
  moving based on the current pending block vs. the latest block
* `latestPriorityFeeRange` — minimum and maximum values for the priority
  fees in the latest block
* `historicalPriorityFeeRange` — minimum and maximum values for priority
  fees over the past 200 blocks
* `priorityFeeTrend` — an indicator for which direction priority fees
  are moving based on the current pending block vs. the latest block