Add caller-specified error for ApprovalController.clear (#656)
`ApprovalController.clear` rejects all pending approvals and clears the state of the controller. This PR adds a "rejection error" parameter to the `clear` method and removes the default error. We decided that there is no reasonable default for this method, since `clear` may be called for a variety of different reasons.