Reduce load on Infura in gas estimate fallback
Usually we use the MetaSwap API for gas fee estimates, but if that is
down for some reason, we fall back to hitting Infura directly.

A little while ago we were working on some UI improvements to the
extension around gas fees, and to support this some changes to the
MetaSwap API were made so that it requests the last 20,000 blocks from
the network and uses their data to obtain a baseline for base fees so
that we can display a network stability/congestion meter.

We made the same changes to the gas fee estimation fallback logic.
**This ended up being a very bad decision.** Because none of the the
requests to Infura that the fallback logic makes are cached, when the
MetaSwap API goes down, the number of requests made to Infura
skyrockets and Infura suffers performance-wise. This problem already
existed, but has been severely exacerbated by the new changes to the
fallback logic. In fact, recently the Infura team had to disable the
`eth_feeHistory` endpoint because some of the Ethereum nodes crashed
under all of the load.

To address this, this commit updates the part of the fallback logic
which is responsible for calculating network congestion to always return
0.5 instead of making the aforementioned 20,000-block request. This
means that the network stability meter in the extension will stay
static, but that is absolutely preferable to blasting Infura.