Remove non-critical gas estimate fallback data
If a request to the Gas API fails for any reason, we attempt to hit
Infura directly and compute the same pieces of data that the Gas API
computes. However, considering that this data supports the new gas fee
UI design in the extension, it is not necessary for us to emulate the
API 100%. We only need data that is critical to allow users to see and
select fees for different priority levels in order to make transactions.
Everything else — a snapshot of recent base and priority fees, whether
the base and priority fees have risen or fallen recently, and how busy
the network is — is not important.

This commit removes these extra computations so as to further limit the
amount of requests that are made to Infura.