Remove leftover .js file
This file appears to have been left over from the TypeScript migration.

Note: This change causes the coverage percentage to drop below our
configured threshold. To address this, we exclude `gas-util.test.ts`
from the coverage report as it has the lowest percentage. We already
plan to backfill these tests in another commit.