Improve support for backfilling types via .d.ts's (#732)
Currently, we exclude test files from TypeScript's purview, so that only
files in `src/` get emitted to `dist/`. This makes sense from a release
perspective, but it interferes with development.

One example of this interference involves [ambient modules][1]. We have
a file `dependencies.d.ts` which we can use to backfill missing types
for various dependencies. The problem is that while TypeScript (or, more
accurately, `tsserver`) can see the types specified here when reading
source files, it doesn't seem to pick them up for test files. This
causes in-editor inconsistencies with type analysis and, in some cases,
warnings such as:

* `import` can only be used with `esInteropModuleFlag: true`
* `X` is missing type definitions

The reason this is happening is that type definition files are not
intended to just be placed in the source directory. You can either use
triple-slash directives to bring them in, or you can modify a setting in
`tsconfig.json` (`paths`) that will [instruct TypeScript where it should
find types for modules that are imported][1]. This commit chooses the
latter option. Note that this requires we split up `dependencies.d.ts`
into separate type definition files, one per module.

In addition to modifying `tsconfig.json`, this commit also breaks up
this config file into a global/development version and a version that is
specific to the build/release process. This allows editors to use
the same exact TypeScript settings for test files as for non-test files.
`tsconfig.build.json` is used by `yarn build` and will limit the files
emitted to `dist/` to only non-test files in `src/`, as in the original
configuration.

[1]: https://www.typescriptlang.org/docs/handbook/modules.html#ambient-modules
[2]: https://www.typescriptlang.org/docs/handbook/module-resolution.html#path-mapping