Update typedoc to v0.22.15
TypeDoc has been updated to the latest version. This version supports
later TypeScript versions, allowing us to update `tsc` in a later PR.
It also includes improvements to document generation, most importantly
in the case where parameters are destructured [1].

Versions v0.21.0 [2] and v0.22.0 [3] both included breaking changes,
but none of them affect us.

This new version included new console warnings that alerted me to a
pre-existing problem that was introduced with v0.20.0, which is that
many types are missing from our documentation. This is because TypeDoc
will only include types that are exported from the package, and many of
our internal types aren't exported from the package itself (even when
they are exported from the module they're declared in). The plugin
`typedoc-plugin-missing-exports` was added to address this. This plugin
ensures that any types referenced in the docs are included in the docs,
even if they aren't exported.

[1]: https://github.com/TypeStrong/typedoc/issues/1703
[2]: https://github.com/TypeStrong/typedoc/releases/tag/v0.21.0
[3]: https://github.com/TypeStrong/typedoc/releases/tag/v0.22.0