Replace WeakMap with private fields (#786)
The `KeyringController` used a WeakMap to keep the keyring as a truly
private field, such that code with a reference to the controller still
could not access it.

This strategy is no longer necessary as of TypeScript 4.3 because we
now have support for true private fields. The TypeScript compiler will
compile this to a similar WeakMap-based strategy.

This should include no functional changes.