Remove unused `retry` parameter from `safelyExecute` (#787)
The `retry` parameter of `safelyExecute` was unused, so it has been
removed. Generally this parameter does not seem useful, and we are
trying to remove all uses of this function anyway.

This makes updating to TypeScript v4.4 slightly easier, because the
invocation of the retry function is a type error on that version
(because `error` is `unknown`).