Remove sinon sandboxes
Sinon uses the "default sandbox" by default, and this is sufficient for
everything we are using Sinon for. People tend to use the
`createSandbox` method because they misunderstand how sandboxes work.
This often results in interwoven use of a module-specific sandbox and
the default sandbox, which leads to errors.

Instead we should use the default sandbox and restore it after each
test.