Update TypeScript to v4.4
See the release announcement for more details about v4.4:
https://devblogs.microsoft.com/typescript/announcing-typescript-4-4-rc/

The main breaking change that affected us was the change in the default
type of the `catch` error parameter from `any` to `unknown`. This was
resolved in some places by changing the expected type to `unknown`. In
cases where we need it to be an `Error` or where we use an `Error`
property, the type is now checked at runtime.

A minor change was needed in `BaseControlelrV2` as well, because
TypeScript wasn't happy with the fact that it was never initialized
for some reason, even though it wasn't supposed to be initialized
(it was set to `never`). But initializing it as `undefined` serves the
exact same purpose and resolved the error.