Move @types/debug to prod dependencies (#23)
In order for another project to use the logging functions provided by
this library, which make use of `debug`, `@types/debug` also needs to be
available. This is not a total blocker — `@types/debug` can be added
manually to the project in question — but is inconvenient. This commit
moves `@types/debug` to the list of `dependencies` to address this.