Add Git hook to run linter before pushing (#898)
It's far too common to create or update a pull request, receive an
approval, and then realize that you can't merge it because CI is saying
that you have linter violations. Ideally, you could always run the
linter before you push, but this step is easy to forget.

To solve this problem, this commit provides a way to install a Git hook
via `simple-git-hooks` which will run `yarn lint` pre-push. If there are
any linter violations, pushing will fail and you can run `yarn lint:fix`
to fix them (or fix them manually).

Note that a previous solution to this problem, submitted as a PR to the
extension repo, involved the use of `lint-staged` to run the linter
against staged files pre-commit. I've changed the workflow to pre-push
because ESLint, when supplied with TypeScript rules, can take a long
time to run (around 2-3 seconds in my testing). This is noticeable
enough to be annoying.

Also note that because the use of Git hooks is local to one's machine,
they must be installed in order to take advantage of automatic linting.
To do this, I've re-introduced `yarn setup`, which will run `yarn
simple-git-hooks`.