Fix bugs in `retryOnEmpty` middleware (#147)
When this package was converted to TypeScript, two bugs were introduced
in the `retryOnEmpty` middleware.

* The goal of this middleware is to retry requests to the network that
return responses whose `result` property is empty in some way, but after
the TypeScript conversion, we were checking whether the whole *response*
object was empty rather than just the `result`. This never happens in
practice, so it ended up making the middleware ineffective.
* Additionally, after retrying the request, the middleware fell through
to the next middleware, which in practice is the middleware that makes a
request to Infura. This meant that later middleware overwrote the
results of the `retryOnEmpty` middleware, also making it ineffective.

This commit restores the functionality prior to the TypeScript
conversion and adds tests to confirm.