Remove the need for an extra tsconfig for tests
Some of our dependencies use ESM and need transpiling in order to use
properly. We previously solved this by creating a special
`tsconfig.json` file for tests where we turned on `allowJs: true`, and
then using the `ts-jest/presets/js-with-ts` Jest preset as well.

We don't need to do this, however, because we can specify
`customExportConditions` instead, which will tell Jest that when it is
resolving imports to ignore `browser` entries in `exports` and choose
`cjs` instead.